requirements.txt
bcrypt==4.0.1
celery==5.4.0
fastapi==0.111.0
sqlalchemy==2.0.30
passlib==1.7.4
phonenumbers==8.13.38
psycopg2-binary==2.9.9
redis==5.0.5
